django/requirements.txt
django>=4.0
django-versionfield>=1.0.2
